pyproject.toml
[tool.poetry]
name = "arc-cli"
version = "0.1.0"
description = ""
authors = ["Lance Chang <virtuouslycan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
typer-cli = "^0.0.12"
shellingham = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
